refactor: prepares exports for deploy
